Revert "Ignore DEP0147"
This reverts commit 2b1de2015d1eb7cc2cafd0860a204e2f7bec90ad.

+++ src/polyfill/fsPoly.ts
@@ -3,6 +3,7 @@ import fs, { PathLike, promises as fsPromises, RmOptions } from 'fs';
 import { isNotJunk } from 'junk';
 import os from 'os';
 import path from 'path';
+import semver from 'semver';
 
 export default class FsPoly {
   /**
@@ -66,8 +67,14 @@ export default class FsPoly {
 
     // Added in: v10.0.0
     if ((await fsPromises.lstat(pathLike)).isDirectory()) {
-      // Added in: v10.0.0
-      await fsPromises.rmdir(pathLike, options);
+      // DEP0147
+      if (semver.lt(process.version, '16.0.0')) {
+        // Added in: v10.0.0
+        await fsPromises.rmdir(pathLike, options);
+      } else {
+        // Added in: v14.14.0
+        await fsPromises.rm(pathLike, { ...options, force: true });
+      }
     } else {
       // Added in: v10.0.0
       await fsPromises.unlink(pathLike);
@@ -90,8 +97,14 @@ export default class FsPoly {
 
     // Added in: v0.1.30
     if (fs.lstatSync(pathLike).isDirectory()) {
-      // Added in: v0.1.21
-      fs.rmdirSync(pathLike, options);
+      // DEP0147
+      if (semver.lt(process.version, '16.0.0')) {
+        // Added in: v0.1.21
+        fs.rmdirSync(pathLike, options);
+      } else {
+        // Added in: v14.14.0
+        fs.rmSync(pathLike, { ...options, force: true });
+      }
     } else {
       // Added in: v0.1.21
       fs.unlinkSync(pathLike);
